+++ R/data.R
@@ -1,0 +1,37 @@
+#' @title A_example dataset
+#'
+#' @description
+#' An example dataset containing artificial personal data.
+#'
+#' @format
+#' A `data.frame` with 10 records. Each row represents one record, with the following columns:
+#' `name`, `surname`, and `city`. Some records can be matched with records in the
+#' `B_example` dataset.
+#'
+#' @docType data
+#' @keywords datasets
+#' @name A_example
+#' @rdname A_example
+#' @examples
+#' data("A_example")
+#' A_example
+"A_example"
+
+#' @title B_example dataset
+#'
+#' @description
+#' An example dataset containing artificial personal data.
+#'
+#' @format
+#' A `data.frame` with 12 records. Each row represents one record, with the following columns:
+#' `name`, `surname`, and `city`. Some records can be matched with records in the
+#' `A_example` dataset.
+#'
+#' @docType data
+#' @keywords datasets
+#' @name B_example
+#' @rdname B_example
+#' @examples
+#' data("B_example")
+#' B_example
+"B_example"