+++ NEWS.md
@@ -10,6 +10,8 @@ based on the following methods: `"binary"`, `"continuous_parametric"`, and `"con
 + Added support for predicting matches based on a record linkage model.
 + Added the unsupervised maximum entropy classification (MEC) algorithm for record linkage.
 Supported methods are: `"binary"`, `"continuous_parametric"`, `"continuous_nonparametric"`, and `"hit-miss"`.
++ Added support for creating the predicted set of matches based on: its estimated size, a target False Link Rate (FLR)
+or a target Missing Match Rate (MMR).
 + Implemented S3 methods for printing.
 + Added support for evaluation when true matches are known.
 + Added documentation and examples.